VBoxVideo: fix integer overflow in parameters check (#5720)
svn:sync-xref-src-repo-rev: r72096

+++ src/VBox/Additions/WINNT/Graphics/Video/disp/xpdm/VBoxDispDriver.cpp
@@ -1,4 +1,4 @@
-/* $Id: VBoxDispDriver.cpp 37273 2011-05-31 11:44:48Z noreply@oracle.com $ */
+/* $Id: VBoxDispDriver.cpp 37335 2011-06-07 08:25:17Z noreply@oracle.com $ */
 
 /** @file
  * VBox XPDM Display driver interface functions
@@ -942,7 +942,7 @@ ULONG APIENTRY VBoxDispDrvEscape(SURFOBJ *pso, ULONG iEsc, ULONG cjIn, PVOID pvI
                 &&  pvIn
                 &&  lpRgnData->rdh.dwSize == sizeof(RGNDATAHEADER)
                 &&  lpRgnData->rdh.iType  == RDH_RECTANGLES
-                &&  cjIn == lpRgnData->rdh.nCount * sizeof(RECT) + sizeof(RGNDATAHEADER))
+                &&  cjIn == ((uint64_t)lpRgnData->rdh.nCount) * sizeof(RECT) + sizeof(RGNDATAHEADER))
             {
                 DWORD   i;
                 PRTRECT pRTRect;
